add minimal_size_type and macro to disable push_back=fast_push_back optimization for static_vector

+++ include/common/memory.hpp
@@ -8,6 +8,7 @@
 
 namespace uwr::mem {
 
+// TODO: update
 /*
  * For now most of functions like
  * construct, fill, ufill, copy,
@@ -23,7 +24,8 @@ namespace uwr::mem {
 
 // TODO: toogle comments
 // using len_t = std::uint_fast32_t;
-using len_t = std::size_t;
+// using len_t = std::size_t;
+using len_t = std::uint_fast64_t;
 
 // should use getpagesize() but want constexpr
 inline constexpr len_t page_size = 4096;
